+++ __tests__/components/Bookmarks.test.js
@@ -5,9 +5,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {assert} from 'chai';
 import raf from 'raf';
-import  ol from 'openlayers';
+import ol from 'openlayers';
 import intl from '../mock-i18n';
-import {Bookmarks} from '../../src/components/Bookmarks';
+import Bookmarks from '../../src/components/Bookmarks';
+import BoundlessSdk from '../../src/components/BoundlessSdk';
+import Map from '../../src/components/Map';
+import configureStore from '../../src/stores/Store';
 
 raf.polyfill();
 
@@ -18,11 +21,15 @@ describe('Bookmarks', function() {
   var bookmarks = [{
     name: 'Le Grenier Pain',
     description: '<b>Address: </b>38 rue des Abbesses<br><b>Telephone:</b> 33 (0)1 46 06 41 81<br><a href="http://www.legrenierapain.com">Website</a>',
-    extent: [259562.7661267497, 6254560.095662868, 260675.9610346824, 6256252.988234103]
+    extent: [259562.7661267497, 6254560.095662868, 260675.9610346824, 6256252.988234103],
+    center: ol.extent.getCenter([259562.7661267497, 6254560.095662868, 260675.9610346824, 6256252.988234103]),
+    zoom: 13
   }, {
     name: 'Poilne',
     description: '<b>Address: </b>8 rue du Cherche-Midi<br><b>Telephone:</b> 33 (0)1 45 48 42 59<br><a href="http://www.poilane.fr">Website</a>',
-    extent: [258703.71361629796, 6248811.5276565505, 259816.90852423065, 6250503.271278702]
+    extent: [258703.71361629796, 6248811.5276565505, 259816.90852423065, 6250503.271278702],
+    center: ol.extent.getCenter([258703.71361629796, 6248811.5276565505, 259816.90852423065, 6250503.271278702]),
+    zoom: 13
   }];
 
   beforeEach(function(done) {
@@ -54,8 +61,13 @@ describe('Bookmarks', function() {
 
   it('zooms to the next bookmark', function(done) {
     var container = document.createElement('div');
+    const store = configureStore();
     ReactDOM.render((
-      <Bookmarks intl={intl} map={map} bookmarks={bookmarks} />
+      <BoundlessSdk store={store}>
+        <Map intl={intl} map={map}>
+          <Bookmarks intl={intl} bookmarks={bookmarks} />
+        </Map>
+      </BoundlessSdk>
     ), container);
     var next = container.querySelector('div.navNext');
     assert.equal(map.getView().getZoom(), 1);
@@ -63,7 +75,7 @@ describe('Bookmarks', function() {
     assert.equal(map.getView().getCenter()[1], 0);
     TestUtils.SimulateNative.click(next);
     window.setTimeout(function() {
-      assert.equal(Math.round(map.getView().getZoom()), 14);
+      assert.equal(Math.round(map.getView().getZoom()), 13);
       assert.equal(map.getView().getCenter()[0], 260119.36358071605);
       assert.equal(map.getView().getCenter()[1], 6255406.541948485);
       ReactDOM.unmountComponentAtNode(container);
@@ -73,8 +85,13 @@ describe('Bookmarks', function() {
 
   it('zooms to the next bookmark using nav dots [1]', function(done) {
     var container = document.createElement('div');
+    const store = configureStore();
     ReactDOM.render((
-      <Bookmarks intl={intl} map={map} bookmarks={bookmarks} />
+      <BoundlessSdk store={store}>
+        <Map intl={intl} map={map}>
+          <Bookmarks intl={intl} map={map} bookmarks={bookmarks} />
+        </Map>
+      </BoundlessSdk>
     ), container);
     var buttonOne = container.querySelector('button#dots_1');
     assert.equal(map.getView().getZoom(), 1);
@@ -82,7 +99,7 @@ describe('Bookmarks', function() {
     assert.equal(map.getView().getCenter()[1], 0);
     TestUtils.SimulateNative.click(buttonOne);
     window.setTimeout(function() {
-      assert.equal(Math.round(map.getView().getZoom()), 14);
+      assert.equal(Math.round(map.getView().getZoom()), 13);
       assert.equal(map.getView().getCenter()[0], 260119.36358071605);
       assert.equal(map.getView().getCenter()[1], 6255406.541948485);
       ReactDOM.unmountComponentAtNode(container);
@@ -92,8 +109,13 @@ describe('Bookmarks', function() {
 
   it('zooms to the next bookmark using nav dots [2]', function(done) {
     var container = document.createElement('div');
+    const store = configureStore();
     ReactDOM.render((
-      <Bookmarks intl={intl} map={map} bookmarks={bookmarks} />
+      <BoundlessSdk store={store}>
+        <Map intl={intl} map={map}>
+          <Bookmarks intl={intl} map={map} bookmarks={bookmarks} />
+        </Map>
+      </BoundlessSdk>
     ), container);
     var buttonTwo = container.querySelector('button#dots_2');
     assert.equal(map.getView().getZoom(), 1);
@@ -101,7 +123,7 @@ describe('Bookmarks', function() {
     assert.equal(map.getView().getCenter()[1], 0);
     TestUtils.SimulateNative.click(buttonTwo);
     window.setTimeout(function() {
-      assert.equal(Math.round(map.getView().getZoom()), 14);
+      assert.equal(Math.round(map.getView().getZoom()), 13);
       assert.equal(map.getView().getCenter()[0], 259260.31107026432);
       assert.equal(map.getView().getCenter()[1], 6249657.399467627);
       ReactDOM.unmountComponentAtNode(container);
